requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
requests==2.31.0
